add HookMixin
