+++ addon/index.js
@@ -1,6 +1,8 @@
 import { hook, $hook } from './test-helpers/hook';
+import HookMixin from './mixins/hook';
 
 export {
   hook,
-  $hook
+  $hook,
+  HookMixin
 };
