--- addon/utils/get-app-config.js
@@ -1,7 +1,0 @@
-export default function getAppConfig() {
-  const configName = Object.keys(requirejs.entries).filter((entry) => {
-    return entry.match(/\/config\/environment/);
-  })[0];
-
-  return requirejs(configName).default;
-}
