--- app/initializers/hook-component-reopener.js
@@ -1,1 +1,0 @@
-export { default, initialize } from 'ember-hook/initializers/hook-component-reopener';
