+++ bower.json
@@ -1,20 +1,20 @@
 {
   "name": "ember-hook",
   "dependencies": {
-    "ember": "2.0.0",
-    "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
+    "ember": "2.1.0",
+    "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.5",
     "ember-cli-test-loader": "ember-cli-test-loader#0.1.3",
-    "ember-data": "2.0.0",
+    "ember-data": "2.1.0",
     "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.5",
-    "ember-qunit": "0.4.9",
+    "ember-qunit": "0.4.10",
     "ember-qunit-notifications": "0.0.7",
-    "ember-resolver": "~0.1.18",
-    "jquery": "^1.11.3",
+    "ember-resolver": "~0.1.21",
+    "jquery": "1.11.3",
     "loader.js": "ember-cli/loader.js#3.2.1",
     "qunit": "~1.18.0"
   },
   "resolutions": {
-    "ember": "2.0.0"
+    "ember": "^2.0.0"
   },
   "devDependencies": {
     "blanket": "~1.1.5"
