+++ tests/dummy/app/components/test-component.js
@@ -1,5 +1,6 @@
 import Ember from 'ember';
+import { HookMixin } from 'ember-hook';
 
 const { Component } = Ember;
 
-export default Component.extend();
+export default Component.extend(HookMixin);
